Running a git commit to attempt to resolve the following:
On branch devel
Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	renamed:    src/git-repo-rewriter/process_json_file.py -> src/git_repo_rewriter/process_json_file.py

As described in the parent commit, devel 21d3922, it appeared that git had recognized this as a deleted file earlier.  Now git reports that the file has been renamed, which appears to be a different response than for the dunder init python file, but as long as it works it should be satisfactory.  In checking this git status, we discovered the dunder init python file is still in place.

If the latest version of the process_json_file.py update is successful in this commit, the changes include defect fixes to resolve an incorrect file path printed in the line 29 `for each_file in subvalue:` processing.  As a result of the most recent changes, the full path to the source files are recognized and printed.  It would be nice to improve the robustness by verifying the path and file actually exist there and handle a file not found or similar error.